requirements.txt
torch
torchvision
pyro-ppl
numpy
pandas
matplotlib
yaml
